qpa: Set minimum device ratio of 1 for windows
This is to prevent rendering issues in the present windows close buttons
as Qt supports scale factors that are less than one poorly.

+++ src/plugins/qpa/window.h
@@ -57,7 +57,7 @@ private:
     EGLSurface m_eglSurface = EGL_NO_SURFACE;
     quint32 m_windowId;
     bool m_resized = false;
-    int m_scale = 1;
+    qreal m_scale = 1;
 };
 
 }
